+++ packages/uikit/src/widgets/DepositModal/index.stories.tsx
@@ -11,8 +11,7 @@ export default {
 export const Wallet: React.FC = () => {
   const { onPresentDepositModal, onPresentDepositModalv2 } = useDepositModal(
     () => null,
-    () => null,
-    "0xbdda50183d817c3289f895a4472eb475967dc980"
+    () => null
   );
   return (
     <Flex>