+++ packages/uikit/src/widgets/NFTModal/useNFTModal.tsx
@@ -1,14 +1,15 @@
 import React from "react";
 import { useModal } from "../Modal";
 import NFTModal from "./NFTModal";
-import NFTInventory from "./nft-inventory";
 
 interface ReturnType {
   onPresentNFTModal: () => void;
 }
 
-const useNFTModal = (login: any): ReturnType => {
-  const [onPresentNFTModal] = useModal(<NFTModal totalNFT={NFTInventory.length} NFTInventory={NFTInventory} />);
+const useNFTModal = (NFTInventory: any, onDismiss: any): ReturnType => {
+  const [onPresentNFTModal] = useModal(
+    <NFTModal totalNFT={NFTInventory.length} NFTInventory={NFTInventory} onDismiss={onDismiss} />
+  );
   return { onPresentNFTModal };
 };