+++ packages/uikit/src/widgets/SellCreditModal/SellCreditModal.tsx
@@ -1,14 +1,13 @@
-import React from 'react'
-import styled from 'styled-components'
-import { Modal } from '../Modal'
-import { Button } from '../../components/Button'
-import './CreateToNature.css'
+import React from "react";
+import styled from "styled-components";
+import { Modal } from "../Modal";
+import { Button } from "../../components/Button";
 
 interface SellCreditModalProps {
-  onDismiss?: () => void
+  onDismiss?: () => void;
 }
 
-const StyledModal = styled(Modal)`
+const StyledModal = styled(Modal)`3
 background-image: url('/images/gradient-603408f80d379f66929884.png');
 padding:4px;
 
@@ -243,7 +242,7 @@ max-width: 320px;
     justify-content: flex-end;
   }
 }
-`
+`;
 
 const SellCreditModal: React.FC<SellCreditModalProps> = ({ onDismiss = () => null }) => {
   return (
@@ -255,35 +254,42 @@ const SellCreditModal: React.FC<SellCreditModalProps> = ({ onDismiss = () => nul
         </div>
       </div>
       <div className="textbox">
-<div className="textbox-content">
-  <div>
-    <div className="textbox-value">
-      <input type="number" className="textbox-input" defaultValue="0"/>
-    </div>
-    <div className="textbox-credit">~0.00 CREDIT</div>
-  </div>
-</div>
+        <div className="textbox-content">
+          <div>
+            <div className="textbox-value">
+              <input type="number" className="textbox-input" defaultValue="0" />
+            </div>
+            <div className="textbox-credit">~0.00 CREDIT</div>
+          </div>
+        </div>
       </div>
       <div className="balance-container">
-<div className="balance-container-nested">
-  <div className="error">Insufficient CREDIT balance</div>
-  <div className="balance">
-    <div className="text-subtle" style={{marginRight: 4, fontSize: 12}}>CREDIT Balance:</div>
-    <div className="text-subtle"style={{ fontSize: 12}}>
-0.000</div>
-  </div>
-</div>
+        <div className="balance-container-nested">
+          <div className="error">Insufficient CREDIT balance</div>
+          <div className="balance">
+            <div className="text-subtle" style={{ marginRight: 4, fontSize: 12 }}>
+              CREDIT Balance:
+            </div>
+            <div className="text-subtle" style={{ fontSize: 12 }}>
+              0.000
+            </div>
+          </div>
+        </div>
       </div>
       <div className="you-pay-container">
         <div className="text-subtle">You sell</div>
-        <div className="text-subtle" style={{color: 'rgb(40, 13, 95)'}}>~0 CREDIT</div>
+        <div className="text-subtle" style={{ color: "rgb(40, 13, 95)" }}>
+          ~0 CREDIT
+        </div>
       </div>
       <div className="enable-container">
-        <Button style={{borderRadius: 'unset'}} onClick={onDismiss}>        Enable
-</Button>
+        <Button style={{ borderRadius: "unset" }} onClick={onDismiss}>
+          {" "}
+          Enable
+        </Button>
       </div>
-  </StyledModal>
-  )
-}
+    </StyledModal>
+  );
+};
 
-export default SellCreditModal
+export default SellCreditModal;